Create IndexNameGenerator utility in server
