+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ilm/GenerateUniqueIndexNameStep.java
@@ -13,16 +13,13 @@ import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.metadata.LifecycleExecutionState;
 import org.elasticsearch.cluster.metadata.LifecycleExecutionState.Builder;
-import org.elasticsearch.cluster.metadata.MetadataCreateIndexService;
-import org.elasticsearch.common.UUIDs;
-import org.elasticsearch.core.Nullable;
 import org.elasticsearch.index.Index;
-import org.elasticsearch.indices.InvalidIndexNameException;
 
-import java.util.Locale;
 import java.util.Objects;
 import java.util.function.BiFunction;
-import java.util.function.Supplier;
+
+import static org.elasticsearch.common.IndexNameGenerator.generateValidIndexName;
+import static org.elasticsearch.common.IndexNameGenerator.validateGeneratedIndexName;
 
 /**
  * Generates a unique index name prefixing the original index name with the configured
@@ -35,8 +32,6 @@ public class GenerateUniqueIndexNameStep extends ClusterStateActionStep {
     private static final Logger logger = LogManager.getLogger(GenerateUniqueIndexNameStep.class);
 
     public static final String NAME = "generate-index-name";
-    static final String ILLEGAL_INDEXNAME_CHARS_REGEX = "[/:\"*?<>|# ,\\\\]+";
-    static final int MAX_GENERATED_UUID_LENGTH = 4;
 
     private final String prefix;
     private final BiFunction<String, Builder, Builder> lifecycleStateSetter;
@@ -98,28 +93,6 @@ public class GenerateUniqueIndexNameStep extends ClusterStateActionStep {
         );
     }
 
-    @Nullable
-    static ActionRequestValidationException validateGeneratedIndexName(String generatedIndexName, ClusterState state) {
-        ActionRequestValidationException err = new ActionRequestValidationException();
-        try {
-            MetadataCreateIndexService.validateIndexOrAliasName(generatedIndexName, InvalidIndexNameException::new);
-        } catch (InvalidIndexNameException e) {
-            err.addValidationError(e.getMessage());
-        }
-        if (state.routingTable().hasIndex(generatedIndexName) || state.metadata().hasIndex(generatedIndexName)) {
-            err.addValidationError("the index name we generated [" + generatedIndexName + "] already exists");
-        }
-        if (state.metadata().hasAlias(generatedIndexName)) {
-            err.addValidationError("the index name we generated [" + generatedIndexName + "] already exists as alias");
-        }
-
-        if (err.validationErrors().size() > 0) {
-            return err;
-        } else {
-            return null;
-        }
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -137,22 +110,4 @@ public class GenerateUniqueIndexNameStep extends ClusterStateActionStep {
         return Objects.hash(super.hashCode(), prefix);
     }
 
-    /**
-     * This generates a valid unique index name by using the provided prefix, appended with a generated UUID, and the index name.
-     */
-    static String generateValidIndexName(String prefix, String indexName) {
-        String randomUUID = generateValidIndexSuffix(UUIDs::randomBase64UUID);
-        randomUUID = randomUUID.substring(0, Math.min(randomUUID.length(), MAX_GENERATED_UUID_LENGTH));
-        return prefix + randomUUID + "-" + indexName;
-    }
-
-    static String generateValidIndexSuffix(Supplier<String> randomGenerator) {
-        String randomSuffix = randomGenerator.get().toLowerCase(Locale.ROOT);
-        randomSuffix = randomSuffix.replaceAll(ILLEGAL_INDEXNAME_CHARS_REGEX, "");
-        if (randomSuffix.length() == 0) {
-            throw new IllegalArgumentException("unable to generate random index name suffix");
-        }
-
-        return randomSuffix;
-    }
 }
